requirements.txt
alembic==1.3.2
datedelta==1.3
email-validator==1.0.5 # Uvicorn complains without this.
psycopg2-binary==2.8.4
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
sentry-sdk==0.14.1
sqlalchemy==1.3.10